refactor tests a bit
